[WS-1] High level review of the code.
TODO's littered throughout the code. Higher-level summary and plan is in
https://github.com/alexmorley/pykilosort/issues/1.
